requirements.txt
absl-py                      2.1.0
albumentations               1.4.4
annotated-types              0.6.0
astunparse                   1.6.3
certifi                      2024.2.2
charset-normalizer           3.3.2
click                        8.1.7
colorama                     0.4.6
contourpy                    1.3.1
cycler                       0.12.1
dlib                         19.24.1
face-recognition             1.3.0
face_recognition_models      0.3.0
filelock                     3.16.1
flatbuffers                  24.3.25
fonttools                    4.55.3
fsspec                       2024.10.0
gast                         0.5.4
google-pasta                 0.2.0
grpcio                       1.62.2
h5py                         3.11.0
idna                         3.7
imageio                      2.34.0
Jinja2                       3.1.4
joblib                       1.4.0
keras                        3.2.1
kiwisolver                   1.4.7
lazy_loader                  0.4
libclang                     18.1.1
Markdown                     3.6
markdown-it-py               3.0.0
MarkupSafe                   2.1.5
matplotlib                   3.10.0
mdurl                        0.1.2
ml-dtypes                    0.3.2
mpmath                       1.3.0
namex                        0.0.8
networkx                     3.3
numpy                        1.26.4
opencv-python                4.10.0.84
opencv-python-headless       4.9.0.80
opt-einsum                   3.3.0
optree                       0.11.0
packaging                    24.0
pandas                       2.2.3
pillow                       10.3.0
pip                          24.0
protobuf                     4.25.3
psutil                       6.1.0
py-cpuinfo                   9.0.0
pydantic                     2.7.0
pydantic_core                2.18.1
Pygments                     2.17.2
pyparsing                    3.2.0
python-dateutil              2.9.0.post0
pytz                         2024.2
PyYAML                       6.0.1
requests                     2.31.0
rich                         13.7.1
scikit-image                 0.23.1
scikit-learn                 1.4.2
scipy                        1.13.0
seaborn                      0.13.2
setuptools                   65.5.0
six                          1.16.0
sympy                        1.13.1
tensorboard                  2.16.2
tensorboard-data-server      0.7.2
tensorflow-intel             2.16.1
tensorflow-io-gcs-filesystem 0.31.0
termcolor                    2.4.0
threadpoolctl                3.4.0
tifffile                     2024.4.18
torch                        2.5.1+cu118
torchaudio                   2.5.1+cu118
torchvision                  0.20.1
tqdm                         4.67.1
typing_extensions            4.11.0
tzdata                       2024.2
ultralytics                  8.3.49
ultralytics-thop             2.0.13
urllib3                      2.2.1
Werkzeug                     3.0.2
wheel                        0.43.0
wrapt                        1.16.0